#DeleteNhanVien

+++ SE104_Group1_TaskManagement/GUI/EmployeesWindow.xaml.cs
@@ -25,23 +25,63 @@ namespace GUI
     /// </summary>
     public partial class EmployeesWindow : Window
     {
-        public static DTO_NhanVien crnNV = new DTO_NhanVien();
+        public static DTO_NhanVien crnUser = new DTO_NhanVien();
         BUS_NhanVien nvManager = new BUS_NhanVien();
         BUS_TaiKhoan tkManager = new BUS_TaiKhoan();
         BindingList<DTO_NhanVien> members = new BindingList<DTO_NhanVien>();
         public EmployeesWindow()
         {
+            
             InitializeComponent();
+            membersDataGrid.LoadingRow += MembersDataGrid_LoadingRow;
+
             this.WindowState = WindowState.Maximized;
-            crnNV = nvManager.GetByID(LoginWindow.crnUser.MANV);
-            if (crnNV.MANV != "") username.Text = crnNV.TENNV;
+
+            setUser();
             var converter = new BrushConverter();
 
 
             //  members.Add(new Employee { ID=1, EmployeeCode = "EP1", EmployeeName = "Nguyen Lam Thanh Triet", Gender="Nam", Technique="Software Engineer", Level="1", Email = "trietn61@gmail.com", Phone = "0985-825-804", Note="None" });
 
             members = nvManager.GetAllData();
+            showMember();
+
+
+        }
+
+        void setUser()
+        {
+            crnUser = nvManager.GetByID(LoginWindow.crnUser.MANV);
+            if (crnUser.MANV != "") username.Text = crnUser.TENNV;
+        }    
+        private void MembersDataGrid_LoadingRow(object? sender, DataGridRowEventArgs e)
+        {
+            var firstCol = membersDataGrid.Columns.FirstOrDefault(c => c.Header.ToString() == "C");
+            e.Row.Loaded += (s, args) =>
+            {
+                var row = (DataGridRow)s;
+                var item = row.Item;
+
+                DTO_NhanVien? nv = item as DTO_NhanVien;
+                if (nv != null && nv.MANV == LoginWindow.crnUser.MANV)
+                {
+                    if (firstCol!= null)
+                    {
+                        var chBx = firstCol.GetCellContent(row) as CheckBox;
+                        if (chBx != null)
+                        {
+                            chBx.Visibility = Visibility.Collapsed;
+                        }
+                    }
+                }
+            };
+        }
+
+       void showMember()
+        {
             membersDataGrid.ItemsSource = members;
+            
+
         }
         private void Add_Button_Click(object sender, RoutedEventArgs e)
         {
@@ -140,8 +180,12 @@ namespace GUI
                         if (res != null && res == true)
                         {
                             members = nvManager.GetAllData();
-                            membersDataGrid.ItemsSource = members;
+                            showMember();
                         }
+                        if (item.MANV == crnUser.MANV)
+                        {
+                            setUser();
+                        }    
                     }
 
                     // Do something with the item...
@@ -150,7 +194,8 @@ namespace GUI
         }
         private void DeleteButton_Click(object sender, RoutedEventArgs e)
         {
-            Button button = sender as Button;
+            
+                Button button = sender as Button;
             if (button != null)
             {
                 // Find the DataGridRow that contains the clicked button
@@ -161,16 +206,20 @@ namespace GUI
                     DTO_NhanVien? item = row.Item as DTO_NhanVien;
                     if (item != null)
                     {
-                        (bool, string) res = nvManager.DeleteByID(item);
-                        if (res.Item1 == true)
+                        MessageBoxResult resu = MessageBox.Show("Bạn đang xóa nhân viên "+item.MANV+", thao tác này không thể quay lại.", "Warning", MessageBoxButton.OKCancel);
+                        if (resu == MessageBoxResult.OK)
                         {
-                            MessageBox.Show(res.Item2);
-                            members = nvManager.GetAllData();
-                            membersDataGrid.ItemsSource = members;
-                        }
-                        else
-                        {
-                            MessageBox.Show(res.Item2);
+                            (bool, string) res = nvManager.DeleteByID(item);
+                            if (res.Item1 == true)
+                            {
+                                MessageBox.Show(res.Item2);
+                                members = nvManager.GetAllData();
+                                showMember();
+                            }
+                            else
+                            {
+                                MessageBox.Show(res.Item2);
+                            }
                         }
 
                     }
@@ -192,6 +241,98 @@ namespace GUI
                 return FindVisualParent<T>(parentObject);
         }
 
-        
+        private void Sel_CheckBox_DataContextChanged(object sender, RoutedEventArgs e)
+        {
+            var chkSelectAll = sender as CheckBox;
+            var firstCol = membersDataGrid.Columns.OfType<DataGridCheckBoxColumn>().FirstOrDefault(c => c.DisplayIndex == 0);
+            if (chkSelectAll == null || firstCol == null || membersDataGrid?.Items == null)
+            {
+                return;
+            }
+            foreach (var item in membersDataGrid.Items)
+            {
+                var chBx = firstCol.GetCellContent(item) as CheckBox;
+                if (chBx == null || chBx.Visibility != Visibility.Visible)
+                {
+                    continue;
+                }
+                chBx.IsChecked = chkSelectAll.IsChecked;
+            }
+        }
+
+        private void Del_Button_Click(object sender, RoutedEventArgs e)
+        {
+            MessageBoxResult res = MessageBox.Show("Bạn đang xóa tất cả nhân viên đã chọn, thao tác này không thể quay lại.", "Warning", MessageBoxButton.OKCancel);
+            if (res == MessageBoxResult.OK)
+            {
+                var firstCol = membersDataGrid.Columns.OfType<DataGridCheckBoxColumn>().FirstOrDefault(c => c.DisplayIndex == 0);
+                if (firstCol == null || membersDataGrid?.Items == null)
+                {
+                    return;
+                }
+                foreach (var item in membersDataGrid.Items)
+                {
+                    var chBx = firstCol.GetCellContent(item) as CheckBox;
+                    if (chBx == null)
+                    {
+                        continue;
+                    }
+                    if (chBx.IsChecked == true)
+                    {
+                        DTO_NhanVien? nv = item as DTO_NhanVien;
+                        if (nv != null)
+                        {
+                            nvManager.DeleteByID(nv);
+                            members = nvManager.GetAllData();
+                            showMember();
+                        }
+                    }
+                }
+
+            }
+        }
+
+        private void membersDataGrid_Loaded(object sender, RoutedEventArgs e)
+        {
+            /*var firstCol = membersDataGrid.Columns.FirstOrDefault(c => c.Header.ToString() == "C");
+            var operationCol = membersDataGrid.Columns.FirstOrDefault(c => c.Header.ToString() == "Operations");
+            foreach (var item in membersDataGrid.Items)
+            {
+                DTO_NhanVien? nv = item as DTO_NhanVien;
+                if (nv != null && nv.MANV == LoginWindow.crnUser.MANV)
+                {
+                    var chBx = firstCol.GetCellContent(item) as CheckBox;
+                    if (chBx == null)
+                    {
+                        continue;
+                    }
+                    else
+                    {
+                        chBx.IsEnabled = false;
+                    }
+                }
+            }*/
+        }
+
+        private void DelButton_Loaded(object sender, RoutedEventArgs e)
+        {
+            Button button = sender as Button;
+            if (button != null)
+            {
+                // Find the DataGridRow that contains the clicked button
+                DataGridRow row = FindVisualParent<DataGridRow>(button);
+                if (row != null)
+                {
+                    // Access the data item behind the row
+                    DTO_NhanVien? item = row.Item as DTO_NhanVien;
+                    if (item != null && item.MANV == LoginWindow.crnUser.MANV)
+                    {
+                        button.Visibility = Visibility.Collapsed; 
+                    }
+                }
+
+                // Do something with the item...
+            }
+        }
     }
 }
